First cut of a two-pass implementation that allows us to walk the search criteria process once and create plans with full knowledge of the criteria, even when applying a pattern to a search term.  Passes all unit tests.  Not yet manually reviewed in detail.  Need to verify optimization to not calculate, aggregate, and sort by relevance scores is present and is applied under the correct conditions.
